implement new allocator
before I was essentially using a bump allocator which is pretty much
impossible to actually use and free up memory when it isnt used. now a
system is setup which makes it possible to free memory for it to be
reused. the actual free function is barely used (yet), and the goal
would be for it to be somehow manually inserted so the programmer does
not have to worry about it.